backend/requirements-dev.txt
-r requirements.txt
mypy>=1.7.0
black>=23.11.0
isort>=4.3.21
autoflake>=1.3.1
flake8>=3.7.9
pytest>=5.4.1
sqlalchemy-stubs>=0.3
pytest-cov>=2.8.1
httpx>=0.24.1